Consolidate `AgentData` into `agentdata.go`
