+++ agent/agentdata/agentdata.go
@@ -5,6 +5,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
+	"os"
 
 	"github.com/multiformats/go-varint"
 	"github.com/storacha/go-ucanto/core/delegation"
@@ -81,4 +82,24 @@ func (ad *AgentData) UnmarshalJSON(b []byte) error {
 	}
 
 	return nil
+}
+
+func (ad AgentData) WriteToFile(path string) error {
+	b, err := json.Marshal(ad)
+	if err != nil {
+		return err
+	}
+
+	return os.WriteFile(path, b, 0600)
+}
+
+func ReadFromFile(path string) (AgentData, error) {
+	b, err := os.ReadFile(path)
+	if err != nil {
+		return AgentData{}, err
+	}
+
+	var ad AgentData
+	json.Unmarshal(b, &ad)
+	return ad, nil
 }
