+++ agent/agentdata/agentdata_test.go
@@ -2,6 +2,7 @@ package agentdata_test
 
 import (
 	"encoding/json"
+	"path"
 	"testing"
 
 	"github.com/storacha/go-ucanto/core/delegation"
@@ -28,6 +29,29 @@ func TestRoundTripAgentData(t *testing.T) {
 
 	var agentDataReturned agentdata.AgentData
 	err = json.Unmarshal(str, &agentDataReturned)
+	require.NoError(t, err)
+
+	require.Equal(t, agentData.Principal, agentDataReturned.Principal)
+	require.Equal(t, delegationsCids(agentData), delegationsCids(agentDataReturned))
+}
+
+func TestWriteReadAgentData(t *testing.T) {
+	dataFilePath := path.Join(t.TempDir(), "agentdata.json")
+
+	agentPrincipal, err := signer.Generate()
+	require.NoError(t, err)
+	del, err := newDelegation()
+	require.NoError(t, err)
+
+	agentData := agentdata.AgentData{
+		Principal:   agentPrincipal,
+		Delegations: []delegation.Delegation{del},
+	}
+
+	err = agentData.WriteToFile(dataFilePath)
+	require.NoError(t, err)
+
+	agentDataReturned, err := agentdata.ReadFromFile(dataFilePath)
 	require.NoError(t, err)
 
 	require.Equal(t, agentData.Principal, agentDataReturned.Principal)
